refactor(core): share one OTLP attribute encoder across logs, metrics and traces
Moves toOtlpAnyValue/toOtlpKeyValueList out of logs into core/utils and widens
them to unknown. Adds a bigint branch, so a bigint attribute encodes as an
int64 rather than falling through to a string.
