+++ .changeset/otlp-bigint-int64.md
@@ -1,0 +1,5 @@
+---
+'@posthog/core': patch
+---
+
+Change `bigint` attributes on logs, metrics and spans to send as an int64 rather than as a string.
