+++ packages/core/src/metrics/index.ts
@@ -10,7 +10,7 @@ import type {
 } from '@posthog/types'
 import type { Logger } from '../types'
 import { isArray, safeSetTimeout } from '../utils'
-import { toOtlpKeyValueList } from '../logs/logs-utils'
+import { toOtlpKeyValueList } from '../utils/otlp-any-value'
 import {
   DEFAULT_HISTOGRAM_BOUNDS,
   bucketIndexFor,
